+++ src/service/service.ts
@@ -217,7 +217,7 @@ export class Service<Definitions extends TaskDefinitions> extends TypedEmitter<S
      * @param callback The callback to run once the stream has been initialized and is ready to consume.
      */
     onStream(callback: OnStreamCallback<typeof this['worker']>) {
-        listenForStream(this.#worker, callback);
+        return listenForStream(this.#worker, callback);
     }
 
     /**